@@ -21,14 +21,14 @@
     }
   },
   "require": {
-    "php": "^7.4|^8.0",
+    "php": "^8.2",
     "ext-json": "*",
-    "guzzlehttp/guzzle": "^7.4"
+    "guzzlehttp/guzzle": "^7.8"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9.5",
-    "friendsofphp/php-cs-fixer": "^3.6",
-    "symfony/var-dumper": "^5.4"
+    "phpunit/phpunit": "^10.5",
+    "friendsofphp/php-cs-fixer": "^3.49",
+    "symfony/var-dumper": "^6.4"
   },
   "autoload-dev": {
     "psr-4": {
