@@ -110,7 +110,7 @@ $rules = [
     'short_scalar_cast' => true,
     'no_unset_cast' => true,
     'no_trailing_comma_in_singleline_array' => true,
-    'single_blank_line_before_namespace' => true,
+    'blank_lines_before_namespace' => true,
 ];
 
 $finder = Finder::create()
